refactor(submodules) replace mongoose submodule
The lastest Mongoose web server has completed changed its interface from old. Although it is nice to have extra features, it is not so convenient to change the way the ReadServer is implemented.

Replace it with older one so that it is consistent with the code that is already used in other part of the project.

Closes  #10