feat(rocksdb): store meta data information in 64 rocksdbs
In ReadServer phase one, all the metadata information (4.7TB in total) are stored in 64 rocksdbs (according to its keys' last three chars). This was to ensure that the rocksdb dees not become too big.

For the demo dataset, although it is small and its metadata information can be stored in single one rocksdb, it would be less trouble to change all other codes. I know this is a hack but it also has some benefits.