fix(webservice): fix a bug that could return duplicated results
This bug is introduced by previous commit in which min_read_length and max_read_length becomes configurable. That also means that a service can config the same figure for max_read_length and min_read_length, suggesting there is only one read length in essence.

In this case, if a user supply a query with exactly the same length as the min_read_length and max_read_length, the service attempts to divide the query into tiles of max_read_length long and min_read_length long, causing duplication.