fix(webservice): fix an invalid characters issue with protobuf message
On the front end server side, the message that was sent to metadata lookup was prefixed with the published topic (the last three chars of the query). The message was stored in a string which uses utf-8.

This change ensures that only copying data from string to zmq::message_t, not from zmq::message_t to string and back which will cause incorrect encoding.