Specify to use proto2 syntax for .proto file
In commit eda5a27, the Protobuf has been moved to v3.1.0. Since v3.0.0,
Protocol Buffers language version 3 (aka proto3) has been introduced:
https://github.com/google/protobuf/releases/tag/v3.0.0

Add syntax = "proto2" to enforce use of older version of ProtoBuf for
which the readserver.proto was designed for. This should complete the
upgrade of ProtoBuf to v3.1.0

Resolves #14.